requirements.txt
Django==3.1.7
black==20.8b1
flake8==3.8.4
jupyter==1.0.0
django-extensions==3.1.1
djangorestframework==3.12.2
drf-extensions==0.7.0